requirements.txt
# requirements.txt
snowflake-connector-python
cffi==1.15.0
pyarrow==6.0.1
pymongo
"pymongo[srv]"
matplotlib
